cloud_functions/CloudStorageLogsManager/requirements.txt
google-cloud-storage
google-cloud-firestore
